knowledge-etl/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "knowledge-etl"
version = "0.1.0"
description = "Book-to-intelligence pipeline for AIOS/Neo knowledge base"
requires-python = ">=3.11"
dependencies = [
    # Core
    "anthropic>=0.46.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    # Extract
    "marker-pdf>=1.10.0",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",
    # Load / Dedup
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    # Retry
    "tenacity>=8.0.0",
]

[project.scripts]
knowledge-etl = "knowledge_etl.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/knowledge_etl"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
